[bugfix/B006_improve-security-configuration-and-message-presentation]: Sessionify the ai-liz-operation microservice. | Adding permission checking into endpoints.

+++ DominionsBackend/servers/ai/liz/server-ai-liz-operation/src/main/java/hu/kalmancheysandor/applications/dominions/servers/ai/liz/operation/controller/LizCharacterController.java
@@ -3,6 +3,8 @@ package hu.kalmancheysandor.applications.dominions.servers.ai.liz.operation.cont
 
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.dto.character.*;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.dto.character.option.LizCharacterVariantOptionResponse;
+import hu.kalmancheysandor.applications.dominions.apis.server.user.admin.AdminPermission;
+import hu.kalmancheysandor.applications.dominions.apis.server.user.admin.service.authentication.AdminAuthenticationService;
 import hu.kalmancheysandor.applications.dominions.servers.ai.liz.operation.service.LizCharacterService;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
@@ -19,11 +21,17 @@ public class LizCharacterController {
     @Autowired
     private LizCharacterService lizCharacterService;
 
+    @Autowired
+    private AdminAuthenticationService authenticationService;
+
     /// ///////////////////////////////////// ACCESS ////////////////////////////////////////////////////////////
 
     @GetMapping("/{uuid}")
     @ResponseStatus(HttpStatus.OK)
     public LizCharacterAccessResponse access(@PathVariable("uuid") String uuid) {
+        // Check permission
+        authenticationService.assertHasAccessPermission(AdminPermission.generator().ai().liz().character());
+
         return lizCharacterService.access(uuid);
     }
 
@@ -31,6 +39,8 @@ public class LizCharacterController {
 
     @GetMapping("/list")
     public List<LizCharacterItemResponse> listAll() {
+        // Check permission
+        authenticationService.assertHasAccessPermission(AdminPermission.generator().ai().liz().character());
 
         return lizCharacterService.listAll();
     }
@@ -39,11 +49,17 @@ public class LizCharacterController {
     @PostMapping("/add")
     @ResponseStatus(HttpStatus.CREATED)
     public LizCharacterCreateResponse add(@Valid @RequestBody LizCharacterCreateRequest request) {
+        // Check permission
+        authenticationService.assertHasAddPermission(AdminPermission.generator().ai().liz().character());
+
         return lizCharacterService.save(request);
     }
 
     @GetMapping("/add/options/variant")
     public List<LizCharacterVariantOptionResponse> atAddListVariant() {
+        // Check permission
+        authenticationService.assertHasAddPermission(AdminPermission.generator().ai().liz().character());
+
         return lizCharacterService.atSaveListVariant();
     }
 
@@ -53,12 +69,16 @@ public class LizCharacterController {
     @PostMapping("/{uuid}/edit")
     @ResponseStatus(HttpStatus.OK)
     public LizCharacterUpdateResponse edit(@PathVariable("uuid") String uuid, @Valid @RequestBody LizCharacterUpdateRequest request) {
+        // Check permission
+        authenticationService.assertHasEditPermission(AdminPermission.generator().ai().liz().character());
+
         return lizCharacterService.update(uuid, request);
     }
 
     @GetMapping("/{uuid}/edit/options/variant")
     public List<LizCharacterVariantOptionResponse> atEditListVariant(@PathVariable("uuid") String uuid) {
-
+        // Check permission
+        authenticationService.assertHasEditPermission(AdminPermission.generator().ai().liz().character());
 
         return lizCharacterService.atUpdateListVariant(uuid);
     }
@@ -68,6 +88,8 @@ public class LizCharacterController {
     @DeleteMapping("/{uuid}/delete")
     @ResponseStatus(HttpStatus.OK)
     public void delete(@PathVariable("uuid") String uuid) {
+        // Check permission
+        authenticationService.assertHasDeletePermission(AdminPermission.generator().ai().liz().character());
 
         lizCharacterService.deleteOneLizCharacter(uuid);
     }
@@ -75,6 +97,9 @@ public class LizCharacterController {
     @DeleteMapping("/delete")
     @ResponseStatus(HttpStatus.OK)
     public void delete(@Valid @RequestBody LizCharacterDeleteRequest request) {
+        // Check permission
+        authenticationService.assertHasDeletePermission(AdminPermission.generator().ai().liz().character());
+
         lizCharacterService.deleteMultiple(request);
     }
