+++ DominionsBackend/servers/ai/liz/server-ai-liz-operation/src/main/java/hu/kalmancheysandor/applications/dominions/servers/ai/liz/operation/controller/LizVariantController.java
@@ -4,6 +4,8 @@ package hu.kalmancheysandor.applications.dominions.servers.ai.liz.operation.cont
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.dto.character.option.LizCharacterVariantOptionResponse;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.dto.variant.*;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.dto.variant.option.LizVariantConceptOptionResponse;
+import hu.kalmancheysandor.applications.dominions.apis.server.user.admin.AdminPermission;
+import hu.kalmancheysandor.applications.dominions.apis.server.user.admin.service.authentication.AdminAuthenticationService;
 import hu.kalmancheysandor.applications.dominions.servers.ai.liz.operation.service.LizVariantService;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
@@ -19,12 +21,17 @@ import java.util.List;
 public class LizVariantController {
     @Autowired
     private LizVariantService lizVariantService;
+    @Autowired
+    private AdminAuthenticationService authenticationService;
 
     /// ///////////////////////////////////// ACCESS ////////////////////////////////////////////////////////////
 
     @GetMapping("/{uuid}")
     @ResponseStatus(HttpStatus.OK)
     public LizVariantAccessResponse access(@PathVariable("uuid") String uuid) {
+        // Check permission
+        authenticationService.assertHasAccessPermission(AdminPermission.generator().ai().liz().variant());
+
         return lizVariantService.access(uuid);
     }
 
@@ -32,6 +39,8 @@ public class LizVariantController {
 
     @GetMapping("/list")
     public List<LizVariantItemResponse> listAll() {
+        // Check permission
+        authenticationService.assertHasAccessPermission(AdminPermission.generator().ai().liz().variant());
 
         return lizVariantService.listAll();
     }
@@ -40,11 +49,17 @@ public class LizVariantController {
     @PostMapping("/add")
     @ResponseStatus(HttpStatus.CREATED)
     public LizVariantCreateResponse add(@Valid @RequestBody LizVariantCreateRequest request) {
+        // Check permission
+        authenticationService.assertHasAddPermission(AdminPermission.generator().ai().liz().variant());
+
         return lizVariantService.save(request);
     }
 
     @GetMapping("/add/options/concept")
     public List<LizVariantConceptOptionResponse> atAddListConcept() {
+        // Check permission
+        authenticationService.assertHasAddPermission(AdminPermission.generator().ai().liz().variant());
+
         return lizVariantService.atSaveListConcept();
     }
 
@@ -53,13 +68,17 @@ public class LizVariantController {
     @PostMapping("/{uuid}/edit")
     @ResponseStatus(HttpStatus.OK)
     public LizVariantUpdateResponse edit(@PathVariable("uuid") String uuid, @Valid @RequestBody LizVariantUpdateRequest request) {
+        // Check permission
+        authenticationService.assertHasEditPermission(AdminPermission.generator().ai().liz().variant());
+
         return lizVariantService.update(uuid, request);
     }
 
 
     @GetMapping("/{uuid}/edit/options/concept")
     public List<LizVariantConceptOptionResponse> atEditListConcept(@PathVariable("uuid") String uuid) {
-
+        // Check permission
+        authenticationService.assertHasEditPermission(AdminPermission.generator().ai().liz().variant());
 
         return lizVariantService.atUpdateListVariant(uuid);
     }
@@ -68,12 +87,18 @@ public class LizVariantController {
     @DeleteMapping("/{uuid}/delete")
     @ResponseStatus(HttpStatus.OK)
     public void delete(@PathVariable("uuid") String uuid) {
+        // Check permission
+        authenticationService.assertHasDeletePermission(AdminPermission.generator().ai().liz().variant());
+
         lizVariantService.deleteOneLizPersonnel(uuid);
     }
 
     @DeleteMapping("/delete")
     @ResponseStatus(HttpStatus.OK)
     public void delete(@Valid @RequestBody LizVariantDeleteRequest request) {
+        // Check permission
+        authenticationService.assertHasDeletePermission(AdminPermission.generator().ai().liz().variant());
+
         lizVariantService.deleteMultiple(request);
     }
